# docs/swagger.yaml
definitions:
  modelsViewApi.BadRequestError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.RaceInfo:
    properties:
      RaceID:
        type: string
      RaceNum:
        type: integer
    type: object
...
  modelsViewApi.RankingResponse:
    properties:
      Races:
        items:
          $ref: '#/definitions/modelsViewApi.RaceInfo'
        type: array
      RankingTable:
        items:
          $ref: '#/definitions/modelsViewApi.RatingTableLine'
        type: array
    type: object
  modelsViewApi.RatingFormData:
    properties:
      blowoutCnt:
        type: integer
      class:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  modelsViewApi.RatingInput:
    properties:
      blowout_cnt:
        type: integer
      class:
        type: integer
      name:
        type: string
    required:
    - class
    - name
    type: object
  modelsViewApi.RatingTableLine:
    properties:
      CoachNames:
        items:
          type: string
        type: array
      ParticipantBirthDates:
        items:
          type: string
        type: array
      ParticipantCategories:
        items:
          type: integer
        type: array
      ParticipantNames:
        items:
          type: string
        type: array
      PointsSum:
        type: integer
      Rank:
        type: integer
      ResInRace:
        additionalProperties:
          type: integer
        type: object
      SailNum:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/ratings:
    get:
      description: Получает список всех рейтингов с возможностью фильтрации
      parameters:
      - description: Фильтр по названию рейтинга
        in: query
        name: name
        type: string
      - description: Фильтр по классу лодки
        in: query
        name: class
        type: string
      - description: Фильтр по количеству гонок, не учитываемых в результате
        in: query
        name: blowoutCnt
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список рейтингов
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.RatingFormData'
            type: array
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
      summary: Получить список всех рейтингов
      tags:
      - Rating
    post:
      consumes:
      - application/json
      parameters:
      - description: Входные данные для создания рейтинга
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.RatingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Рейтинг успешно создан
          schema:
            $ref: '#/definitions/modelsViewApi.RatingInput'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
      summary: Создать новый рейтинг
      tags:
      - Rating
  /api/ratings/{ratingID}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга, который нужно удалить
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      responses:
        "204":
          description: Рейтинг успешно удалён
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг не найден
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить рейтинг
      tags:
      - Rating
    get:
      consumes:
      - application/json
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о рейтинге успешно получена
          schema:
            $ref: '#/definitions/modelsViewApi.RatingFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг не найден
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию о рейтинге
      tags:
      - Rating
    put:
      consumes:
      - application/json
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Входные данные для обновления рейтинга
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.RatingInput'
      produces:
      - application/json
      responses:
        "200":
          description: Рейтинг успешно обновлен
          schema:
            $ref: '#/definitions/modelsViewApi.RatingFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Обновить информацию о рейтинге
      tags:
      - Rating
...
  /api/ratings/{ratingID}/rankings:
    get:
      parameters:
      - description: ID рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      responses:
        "200":
          description: Рейтинговая таблица успешно получена
          schema:
            $ref: '#/definitions/modelsViewApi.RankingResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг не найден
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить рейтинговую таблицу
      tags:
      - Ranking
swagger: "2.0"